backend/app/schemas/feedback.py
from datetime import datetime

from pydantic import BaseModel, ConfigDict


class FeedbackCreate(BaseModel):
    message_id: int
    is_positive: bool
    comment: str | None = None


class FeedbackRead(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: int
    message_id: int
    user_id: int
    is_positive: bool
    comment: str | None = None
    created_at: datetime
