backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
pgvector==0.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.8.1
email-validator==2.2.0
langchain==0.3.19
langgraph==0.2.74
langchain-groq==0.2.4
langchain-community==0.3.18
langchain-text-splitters==0.3.6
duckduckgo-search==7.5.2
beautifulsoup4==4.13.3
requests==2.32.3
python-docx==1.1.2
docx2txt==0.8
pypdf==5.3.0
markdown==3.7
scikit-learn==1.5.2
pytest==8.3.5
httpx==0.28.1
